scraping/requirements.txt
praw==7.8.1
requests==2.32.3
yt-dlp==2026.3.17
youtube-transcript-api==0.6.3
python-dotenv==1.0.1
tqdm==4.67.1
